Added test for PeekIterator, moving it to separate module.
